Prevent empty strings when adding event
DynamoDB won't accept empty strings thus the server will return a GraphQL error as a result in this case. Since GraphQL errors are currently not considered in the result handler and there's no convenient way to revert the optimistically updated local cache, the most straight-forward way to deal with this problem is to notify the user.